separate test on binary field (#1983)
